requirements.txt
# core libs
torch
transformers
datasets
faiss-cpu
Pillow
gradio
tqdm

